# Milestone 4/backend/swagger.yaml
openapi: 3.1.0
info:
  title: EduAssist API
  description: AI-Powered Teaching Assistant Platform - Knowledge Assistant Feature
  version: 1.0.0
paths:
...
  /study-guides/study-guides/video-info:
    get:
      tags:
        - Study Guides
        - Study Guide Generator
      summary: Get Video Info
      description: |-
        Get YouTube video information

        Check if video has captions and get metadata before creating study guide
      operationId: get_video_info_study_guides_study_guides_video_info_get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
            title: Url
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /study-guides/study-guides/:
    post:
      tags:
        - Study Guides
        - Study Guide Generator
      summary: Create Study Guide
      description: |-
        Create a study guide from YouTube video

        Steps:
        1. Validate YouTube URL and get video ID
        2. Get video metadata
        3. Create study guide record (PROCESSING status)
        4. Generate content in background (async)
        5. Return study guide ID immediately

        The client should poll /study-guides/{id} to check status
      operationId: create_study_guide_study_guides_study_guides__post
      security:
        - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudyGuideCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyGuideResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - Study Guides
        - Study Guide Generator
      summary: List Study Guides
      description: List study guides
      operationId: list_study_guides_study_guides_study_guides__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: course_id
          in: query
          required: false
          schema:
            type: integer
            title: Course Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyGuideListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /study-guides/study-guides/{guide_id}:
    get:
      tags:
        - Study Guides
        - Study Guide Generator
      summary: Get Study Guide
      description: Get study guide by ID
      operationId: get_study_guide_study_guides_study_guides__guide_id__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: guide_id
          in: path
          required: true
          schema:
            type: integer
            title: Guide Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyGuideResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - Study Guides
        - Study Guide Generator
      summary: Delete Study Guide
      description: Delete study guide
      operationId: delete_study_guide_study_guides_study_guides__guide_id__delete
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: guide_id
          in: path
          required: true
          schema:
            type: integer
            title: Guide Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /study-guides/study-guides/transcript/{video_id}:
    get:
      tags:
        - Study Guides
        - Study Guide Generator
      summary: Get Video Transcript
      description: |-
        Get video transcript for preview

        Used by frontend to show transcript before creating study guide
      operationId: get_video_transcript_study_guides_study_guides_transcript__video_id__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: video_id
          in: path
          required: true
          schema:
            type: string
            title: Video Id
        - name: start
          in: query
          required: false
          schema:
            type: integer
            default: 0
            title: Start
        - name: end
          in: query
          required: false
          schema:
            type: integer
            title: End
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /slide-decks/slide-decks/:
    post:
      tags:
        - Slide Decks
        - Slide Decks
      summary: Start Slide Deck
      description: Generate slide deck from uploaded docs. Returns deck_id (status PENDING/COMPLETE)
      operationId: start_slide_deck_slide_decks_slide_decks__post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SlideDeckCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Start Slide Deck Slide Decks Slide Decks  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - Slide Decks
        - Slide Decks
      summary: List Decks
      operationId: list_decks_slide_decks_slide_decks__get
      parameters:
        - name: status
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SlideDeckOut'
                title: Response List Decks Slide Decks Slide Decks  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /slide-decks/slide-decks/{deck_id}:
    get:
      tags:
        - Slide Decks
        - Slide Decks
      summary: Get Deck
      operationId: get_deck_slide_decks_slide_decks__deck_id__get
      parameters:
        - name: deck_id
          in: path
          required: true
          schema:
            type: integer
            title: Deck Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlideDeckOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /slide-decks/slide-decks/{deck_id}/slide/{slide_id}:
    put:
      tags:
        - Slide Decks
        - Slide Decks
      summary: Update Slide
      operationId: update_slide_slide_decks_slide_decks__deck_id__slide__slide_id__put
      parameters:
        - name: deck_id
          in: path
          required: true
          schema:
            type: integer
            title: Deck Id
        - name: slide_id
          in: path
          required: true
          schema:
            type: integer
            title: Slide Id
        - name: content_md
          in: query
          required: true
          schema:
            type: string
            title: Content Md
        - name: notes_md
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Notes Md
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlideOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Priority:
      type: string
      enum:
        - high
        - medium
        - low
      title: Priority
    PrioritySegment:
      properties:
        start:
          type: integer
          minimum: 0
          title: Start
          description: Start time in seconds
        end:
          type: integer
          exclusiveMinimum: 0
          title: End
          description: End time in seconds
        priority:
          $ref: '#/components/schemas/Priority'
          default: medium
      type: object
      required:
        - start
        - end
      title: PrioritySegment
      description: User-selected important segment
...
    SlideDeckControls:
      properties:
        num_slides:
          type: integer
          maximum: 30
          minimum: 3
          title: Num Slides
          default: 10
        desired_depth:
          type: string
          title: Desired Depth
          description: summary, normal, in-depth
          default: summary
      type: object
      title: SlideDeckControls
    SlideDeckCreate:
      properties:
        title:
          type: string
          title: Title
        controls:
          $ref: '#/components/schemas/SlideDeckControls'
        reference_document_ids:
          anyOf:
            - items:
                type: integer
              type: array
            - type: 'null'
          title: Reference Document Ids
          default: []
      type: object
      required:
        - title
        - controls
      title: SlideDeckCreate
    SlideDeckOut:
      properties:
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        status:
          $ref: '#/components/schemas/SlideDeckStatus'
        controls:
          $ref: '#/components/schemas/SlideDeckControls'
        slides:
          items:
            $ref: '#/components/schemas/SlideOut'
          type: array
          title: Slides
          default: []
        created_at:
          type: string
          title: Created At
      type: object
      required:
        - id
        - title
        - status
        - controls
        - created_at
      title: SlideDeckOut
    SlideDeckStatus:
      type: string
      enum:
        - PENDING
        - GENERATING
        - COMPLETE
        - FAILED
      title: SlideDeckStatus
    SlideOut:
      properties:
        title:
          type: string
          title: Title
        content_md:
          type: string
          title: Content Md
        notes_md:
          anyOf:
            - type: string
            - type: 'null'
          title: Notes Md
        image_url:
          anyOf:
            - type: string
            - type: 'null'
          title: Image Url
        position:
          type: integer
          title: Position
        id:
          type: integer
          title: Id
      type: object
      required:
        - title
        - content_md
        - position
        - id
      title: SlideOut
    StudyGuideCreate:
      properties:
        youtube_url:
          type: string
          minLength: 1
          title: Youtube Url
        course_id:
          type: integer
          title: Course Id
        title:
          type: string
          maxLength: 255
          minLength: 1
          title: Title
        priority_segments:
          items:
            $ref: '#/components/schemas/PrioritySegment'
          type: array
          minItems: 1
          title: Priority Segments
      type: object
      required:
        - youtube_url
        - course_id
        - title
        - priority_segments
      title: StudyGuideCreate
      description: Request to create study guide from YouTube
    StudyGuideListResponse:
      properties:
        guides:
          items:
            $ref: '#/components/schemas/StudyGuideResponse'
          type: array
          title: Guides
        total:
          type: integer
          title: Total
      type: object
      required:
        - guides
        - total
      title: StudyGuideListResponse
      description: List of study guides
    StudyGuideResponse:
      properties:
        id:
          type: integer
          title: Id
        youtube_url:
          type: string
          title: Youtube Url
        video_id:
          type: string
          title: Video Id
        video_title:
          anyOf:
            - type: string
            - type: 'null'
          title: Video Title
        video_duration:
          anyOf:
            - type: integer
            - type: 'null'
          title: Video Duration
        title:
          type: string
          title: Title
        content:
          anyOf:
            - type: string
            - type: 'null'
          title: Content
        key_topics:
          anyOf:
            - items:
                type: string
              type: array
            - type: 'null'
          title: Key Topics
        status:
          $ref: '#/components/schemas/StudyGuideStatus'
        course_id:
          type: integer
          title: Course Id
        created_by:
          type: integer
          title: Created By
        created_at:
          type: string
          format: date-time
          title: Created At
        segments:
          items:
            $ref: '#/components/schemas/VideoSegmentResponse'
          type: array
          title: Segments
          default: []
      type: object
      required:
        - id
        - youtube_url
        - video_id
        - title
        - status
        - course_id
        - created_by
        - created_at
      title: StudyGuideResponse
      description: Study guide response
    StudyGuideStatus:
      type: string
      enum:
        - PROCESSING
        - COMPLETED
        - FAILED
      title: StudyGuideStatus
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
    VideoSegmentResponse:
      properties:
        id:
          type: integer
          title: Id
        start_time:
          type: integer
          title: Start Time
        end_time:
          type: integer
          title: End Time
        transcript:
          type: string
          title: Transcript
        priority:
          type: string
          title: Priority
        summary:
          anyOf:
            - type: string
            - type: 'null'
          title: Summary
        key_points:
          anyOf:
            - items:
                type: string
              type: array
            - type: 'null'
          title: Key Points
      type: object
      required:
        - id
        - start_time
        - end_time
        - transcript
        - priority
      title: VideoSegmentResponse
      description: Video segment with transcript
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: token